no call_canister module in station api
